Refactor Event component
