+++ lib/ui-components/Event/MessageContainer.jsx
@@ -1,0 +1,77 @@
+/*
+ * Copyright (C) Balena.io - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited.
+ * Proprietary and confidential.
+ */
+
+import {
+	Box
+}	from 'rendition'
+import styled from 'styled-components'
+
+const MessageContainer = styled(Box) `
+	border-radius: 6px;
+	border-top-left-radius: 0;
+	box-shadow: -5px 4.5px 10.5px 0 rgba(152, 173, 227, 0.08);
+
+	a {
+		color: inherit;
+		text-decoration: underline;
+	}
+
+	a .rendition-tag--personal,
+	.rendition-tag--personal {
+		background: #FFF1C2;
+		color: #333;
+
+		&.rendition-tag--read:after {
+			background: #FFC19B;
+			border-radius: 5px;
+			font-size: 8px;
+			padding: 2px;
+		}
+	}
+
+	img {
+		background-color: transparent !important;
+
+		&.emoji {
+			width: 20px;
+			height: 20px;
+			vertical-align: middle;
+		}
+	}
+
+	code {
+		color: #333;
+		background-color: #f6f8fa;
+	}
+
+	${({
+		card, actor, theme
+	}) => {
+		if (card.type === 'whisper' || card.type === 'whisper@1.0.0') {
+			return `
+				background: ${theme.colors.secondary.main};
+				color: white;
+
+				blockquote {
+					color: lightgray;
+				}
+			`
+		}
+
+		if (actor && actor.proxy) {
+			return `
+				background: ${theme.colors.quartenary.main};
+				color: ${theme.colors.text.dark};
+			`
+		}
+
+		return `
+			border: solid 0.5px #e8ebf2;
+			background: white;
+		`
+	}}
+`
+export default MessageContainer
